Add files via upload
In this analysis, an unbalanced 3-way analysis of variance was conducted in order to analyze the effects of 3 different factors on plants’ resistance to cold weather. These factors included the species of potato (1 & 2), the temperature they
were previously acclimatized to (cold & room temperature), and the temperature they were later exposed to (-4 degrees celsius & -8 degrees celsius). After discovery of an insignificant 3-way interaction term, 3 main effects were discovered to be significant at the 95% family confidence level, but due to the interactions between the factors, the interpretation of these main effects were limited to specific treatment combinations.